+++ CHANGELOG.md
@@ -4,15 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [2.1.1] - 2023-09-11
+## [2.2.0] - 2023-09-14
 
 ### Added
 
-- Added new `h-dropdown__menu` class
+- Added new `TextInput` component
+- Added new `Label` component
+- Added new `RadioButton` component
+- Added new `ProgressBar` component
+- Added specs for `TextInput`, `Label`, `RadioButton`, and `ProgressBar`
+- Added new `_progress_bar.scss` file
+- Added new `_radio_button.scss` file
+- Added new `_flex.scss` file
+- Added new `_position.scss` file
 
 ### Changed
 
-- Update `Dropdown` component to use `h-dropdown__menu` class
+- Updated `_forms.scss` file to include styling for new components
+- Updated `_color.scss` variables file to include teal-500 color
+- Updated `_component.scss` variables file to include more input UI variables
 
 ## [2.1.0] - 2023-09-07